Move countdownSeconds into a common helper
The desktop and mobile entry bars each carried an identical private
copy of the round/cast-time countdown composable; both now use one
internal helper in commonMain.
