+++ compose/src/mobileMain/kotlin/warlockfe/warlock3/compose/ui/game/WarlockEntry.kt
@@ -266,27 +266,6 @@ fun BoxScope.RoundTimeBar(
     }
 }
 
-@Composable
-private fun countdownSeconds(
-    endTime: Instant?,
-    getCurrentTime: () -> Instant,
-): Int {
-    var seconds by remember { mutableIntStateOf(0) }
-    val currentGetCurrentTime by rememberUpdatedState(getCurrentTime)
-    LaunchedEffect(endTime) {
-        while (endTime != null) {
-            val now = currentGetCurrentTime()
-            val remaining = endTime - now
-            val remainingMs = remaining.inWholeMilliseconds
-            seconds = ((remainingMs + 999) / 1000).toInt()
-            if (remaining < Duration.ZERO) break
-            val msUntilNextTick = remainingMs % 1000
-            delay((if (msUntilNextTick == 0L) 1000L else msUntilNextTick).milliseconds)
-        }
-    }
-    return seconds
-}
-
 @Preview(widthDp = 800, backgroundColor = 0xFF444444)
 @Composable
 private fun WarlockEntryDarkPreview() {
